+++ app/assets/javascripts/Components/markus_file_manager.jsx
@@ -21,6 +21,11 @@ class RawFileManager extends RawFileBrowser {
     this.handleActionBarAddFolderClick(event);
   };
 
+  handleActionBarSubmitURLClick = (event, selectedItem) => {
+    event.preventDefault();
+    this.props.onActionBarSubmitURLClick(this.folderTarget(selectedItem));
+  };
+
   folderTarget = selectedItem => {
     // treat multiple selections as not targeting a folder
     const selectionIsFolder = !!selectedItem && selectedItem.relativeKey.endsWith("/");
@@ -129,6 +134,20 @@ class RawFileManager extends RawFileBrowser {
             </li>
           );
         }
+        if (this.props.enableUrlSubmit) {
+          actions.unshift(
+            <li key="action-add-link">
+              <a
+                onClick={event => this.handleActionBarSubmitURLClick(event, selectedItem)}
+                href="#"
+                role="button"
+              >
+                <i className="fa fa-submit-link-o" aria-hidden="true" />
+                &nbsp;{I18n.t("submit_the", {item: I18n.t("submissions.student.link")})}
+              </a>
+            </li>
+          );
+        }
         // NEW
         actions.unshift(
           <li key="action-add-file>">
@@ -138,7 +157,7 @@ class RawFileManager extends RawFileBrowser {
               role="button"
             >
               <i className="fa fa-add-file-o" aria-hidden="true" />
-              &nbsp;{I18n.t("upload_the", {item: I18n.t("file")})}
+              &nbsp;{I18n.t("submit_the", {item: I18n.t("file")})}
             </a>
           </li>
         );
@@ -159,6 +178,20 @@ class RawFileManager extends RawFileBrowser {
           </li>
         );
       }
+      if (this.props.enableUrlSubmit) {
+        actions.unshift(
+          <li key="action-add-link">
+            <a
+              onClick={event => this.handleActionBarSubmitURLClick(event, selectedItem)}
+              href="#"
+              role="button"
+            >
+              <i className="fa fa-submit-link-o" aria-hidden="true" />
+              &nbsp;{I18n.t("submit_the", {item: I18n.t("submissions.student.link")})}
+            </a>
+          </li>
+        );
+      }
       // NEW
       actions.unshift(
         <li key="action-add-file>">
@@ -168,7 +201,7 @@ class RawFileManager extends RawFileBrowser {
             role="button"
           >
             <i className="fa fa-add-file-o" aria-hidden="true" />
-            &nbsp;{I18n.t("upload_the", {item: I18n.t("file")})}
+            &nbsp;{I18n.t("submit_the", {item: I18n.t("file")})}
           </a>
         </li>
       );