+++ app/assets/javascripts/Components/repo_browser.jsx
@@ -111,6 +111,7 @@ class RepoBrowser extends React.Component {
           onChange={this.fetchRevisions}
           fetchOnMount={false}
           enableSubdirs={this.props.enableSubdirs}
+          enableUrlSubmit={this.props.enableUrlSubmit}
           readOnly={this.isReadOnly()}
         />
         <ManualCollectionForm