+++ db/migrate/20220128182721_add_url_submit_to_assignment_properties.rb
@@ -1,0 +1,5 @@
+class AddUrlSubmitToAssignmentProperties < ActiveRecord::Migration[6.1]
+  def change
+    add_column :assignment_properties, :url_submit, :boolean, default: false, null: false
+  end
+end