+++ app/helpers/automated_tests_helper.rb
@@ -107,31 +107,6 @@ module AutomatedTestsHelper
     end
   end
 
-  def run_autotester_command(command, server_kwargs)
-    server_username = Settings.autotest.server_username
-    server_command = Settings.autotest.server_command
-    output = ''
-    if server_username.nil?
-      # local cancellation with no authentication
-      args = [server_command, command, '-j', JSON.generate(server_kwargs)]
-      output, status = Open3.capture2e(*args)
-      if status.exitstatus != 0
-        raise output
-      end
-    else
-      # local or remote cancellation with authentication
-      server_host = Settings.autotest.server_host
-      Net::SSH.start(server_host, server_username, auth_methods: ['publickey']) do |ssh|
-        args = "#{server_command} #{command} -j '#{JSON.generate(server_kwargs)}'"
-        output = ssh.exec!(args)
-        if output.exitstatus != 0
-          raise output
-        end
-      end
-    end
-    output
-  end
-
   # Sends RESTful api requests to the autotester
   module AutotestApi
     AUTOTEST_USERNAME = "markus_#{Rails.application.config.action_controller.relative_url_root}".freeze