+++ Changelog.md
@@ -1,10 +1,13 @@
 # Changelog
 
 ## [unreleased]
+- Remove group name displayed attribute from assignment properties table (#5834)
+- Fix bug for menu icon not working on mobile devices / smaller screens (#5818)
 - Remove unmaintained locales (#5727)
 - Fix bug for "Delete Group" button generating an invalid path (#5768)
 - Update wiki urls to point to https://github.com/MarkUsProject/Wiki (#5781)
 - Added the ability to submit URLs (#5822)
+- Allow TAs to download grades for students they've been assigned to grade (#5793)
 
 ## [v2.0.5]
 - Add ability to annotate notebook (jupyter and Rmd) submissions (#5749)