+++ Changelog.md
@@ -6,6 +6,8 @@
 - Remove unmaintained locales (#5727)
 - Fix bug for "Delete Group" button generating an invalid path (#5768)
 - Update wiki urls to point to https://github.com/MarkUsProject/Wiki (#5781)
+- Added the ability to submit URLs (#5822)
+- Allow TAs to download grades for students they've been assigned to grade (#5793)
 
 ## [v2.0.5]
 - Add ability to annotate notebook (jupyter and Rmd) submissions (#5749)